+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- **Breaking** generic: `ProjectFile::get_path`も戻り値が正しくNoneになるように
 - generic: 複数の種類のメニューを登録できるようにする
 - alias: iter_subtables_as_array_mutを追加